+++ lib/hsa/CMakeLists.txt
@@ -10,7 +10,7 @@ endif (HAS_HSA EQUAL 1)
 ####################
 
 if (HAS_HSA EQUAL 1)
-add_mcwamp_library_hsa(mcwamp_hsa mcwamp_hsa.cpp staging_buffer.cpp ../md5.cpp)
+add_mcwamp_library_hsa(mcwamp_hsa mcwamp_hsa.cpp unpinned_copy_engine.cpp ../md5.cpp)
 add_mcwamp_library_hsa(hc_am hc_am.cpp)
 install(TARGETS mcwamp_hsa hc_am
         LIBRARY DESTINATION lib
